@@ -4,14 +4,14 @@ version: 0.0.1
 publish_to: "none"
 
 environment:
-  sdk: ^3.5.2
-  flutter: ^3.29.2
+  sdk: ^3.8.0
+  flutter: ^3.32.2
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_animate: ^4.5.0
-  flutter_svg: ^2.0.10+1
+  flutter_animate: ^4.5.2
+  flutter_svg: ^2.2.0
   mocktail: ^1.0.4
   zup_core:
     # path: ../zup_core
